Modeltime Fit Resample Helpers

Used for low-level resample fitting of modeltime, parnsip and workflow models These functions are not intended for user use.

Usage

mdl_time_fit_resamples(object, resamples, control = control_resamples())

Arguments

object

A Modeltime Table

resamples

An rset resample object. Used to generate sub-model predictions for the meta-learner. See timetk::time_series_cv() or rsample::vfold_cv() for making resamples.

control

A tune::control_resamples() object to provide control over the resampling process.

Value

A tibble with forecast features