Calculate Accuracy Metrics from Modeltime Resamples

This is a wrapper for yardstick that simplifies time series regression accuracy metric calculations from a Modeltime Table that has been resampled and fitted using modeltime_fit_resamples().

Usage

modeltime_resample_accuracy(
  object,
  summary_fns = mean,
  metric_set = default_forecast_accuracy_metric_set(),
  ...
)

Arguments

object

a Modeltime Table with a column '.resample_results' (the output of modeltime_fit_resamples())

summary_fns

One or more functions to analyze resamples. The default is mean(). Possible values are:

NULL, to returns the resamples untransformed.

A function, e.g. mean.

A purrr-style lambda, e.g. ~ mean(.x, na.rm = TRUE)

A list of functions/lambdas, e.g. list(mean = mean, sd = sd)

metric_set

A yardstick::metric_set() that is used to summarize one or more forecast accuracy (regression) metrics.

...

Additional arguments passed to the function calls in summary_fns.

Details

#' Default Accuracy Metrics

The following accuracy metrics are included by default via modeltime::default_forecast_accuracy_metric_set():

MAE - Mean absolute error, yardstick::mae()

MAPE - Mean absolute percentage error, yardstick::mape()

MASE - Mean absolute scaled error, yardstick::mase()

SMAPE - Symmetric mean absolute percentage error, yardstick::smape()

RMSE - Root mean squared error, yardstick::rmse()

RSQ - R-squared, yardstick::rsq()

Summary Functions

By default, modeltime_resample_accuracy() returns the average accuracy metrics for each resample prediction.

The user can change this default behavior using summary_fns. Simply pass one or more Summary Functions. Internally, the functions are passed to dplyr::across(.fns), which applies the summary functions.

Returning Unsummarized Results

You can pass summary_fns = NULL to return unsummarized results by .resample_id.

Professional Tables (Interactive & Static)

Use modeltime::table_modeltime_accuracy() to format the results for reporting in reactable (interactive) or gt (static) formats, which are perfect for Shiny Apps (interactive) and PDF Reports (static).

Examples

library(modeltime)

# Mean (Default)
m750_training_resamples_fitted %>%
    modeltime_resample_accuracy() %>%
    table_modeltime_accuracy(.interactive = FALSE)

  

Accuracy Table
    

.model_id
      	.model_desc
      	.type
      	n
      	mae
      	mape
      	mase
      	smape
      	rmse
      	rsq
    

1
	ARIMA(0,1,1)(0,1,1)[12]
	Resamples
	6
	294.51
	2.95
	1.18
	2.94
	355.49
	0.80

2
	PROPHET
	Resamples
	6
	458.05
	4.59
	1.85
	4.55
	515.49
	0.78

3
	GLMNET
	Resamples
	6
	706.10
	7.16
	2.80
	6.92
	753.04
	0.77

# Mean and Standard Deviation
m750_training_resamples_fitted %>%
    modeltime_resample_accuracy(
        summary_fns = list(mean = mean, sd = sd)
    ) %>%
    table_modeltime_accuracy(.interactive = FALSE)

  

Accuracy Table
    

.model_id
      	.model_desc
      	.type
      	n
      	mae_mean
      	mae_sd
      	mape_mean
      	mape_sd
      	mase_mean
      	mase_sd
      	smape_mean
      	smape_sd
      	rmse_mean
      	rmse_sd
      	rsq_mean
      	rsq_sd
    

1
	ARIMA(0,1,1)(0,1,1)[12]
	Resamples
	6
	294.51
	121.98
	2.95
	1.27
	1.18
	0.56
	2.94
	1.24
	355.49
	138.68
	0.80
	0.15

2
	PROPHET
	Resamples
	6
	458.05
	196.09
	4.59
	1.98
	1.85
	0.91
	4.55
	1.86
	515.49
	175.26
	0.78
	0.09

3
	GLMNET
	Resamples
	6
	706.10
	248.29
	7.16
	2.62
	2.80
	1.11
	6.92
	2.41
	753.04
	252.75
	0.77
	0.16

# When summary_fns = NULL, returns the unsummarized resample results
m750_training_resamples_fitted %>%
    modeltime_resample_accuracy(
        summary_fns = NULL
    )
#> # A tibble: 18 × 10
#>    .model_id .model_desc  .resample_id .type   mae  mape  mase smape  rmse   rsq
#>        <int> <chr>        <chr>        <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1         1 ARIMA(0,1,1… Slice1       Resa…  182.  1.79 0.637  1.80  212. 0.950
#>  2         1 ARIMA(0,1,1… Slice2       Resa…  330.  3.24 1.30   3.17  406. 0.729
#>  3         1 ARIMA(0,1,1… Slice3       Resa…  359.  3.57 1.55   3.64  415. 0.859
#>  4         1 ARIMA(0,1,1… Slice4       Resa…  456.  4.76 2.01   4.61  530. 0.561
#>  5         1 ARIMA(0,1,1… Slice5       Resa…  122.  1.23 0.524  1.22  165. 0.926
#>  6         1 ARIMA(0,1,1… Slice6       Resa…  317.  3.12 1.05   3.20  407. 0.763
#>  7         2 PROPHET      Slice1       Resa…  450.  4.54 1.57   4.40  509. 0.869
#>  8         2 PROPHET      Slice2       Resa…  449.  4.34 1.76   4.45  494. 0.745
#>  9         2 PROPHET      Slice3       Resa…  475.  4.72 2.05   4.85  556. 0.660
#> 10         2 PROPHET      Slice4       Resa…  317.  3.33 1.39   3.26  362. 0.781
#> 11         2 PROPHET      Slice5       Resa…  813.  8.22 3.49   7.89  829. 0.897
#> 12         2 PROPHET      Slice6       Resa…  244.  2.40 0.808  2.45  344. 0.755
#> 13         3 GLMNET       Slice1       Resa…  893.  8.90 3.12   8.48  926. 0.845
#> 14         3 GLMNET       Slice2       Resa…  583.  5.73 2.29   5.53  647. 0.654
#> 15         3 GLMNET       Slice3       Resa…  874.  9.07 3.77   8.62  917. 0.901
#> 16         3 GLMNET       Slice4       Resa…  964. 10.1  4.24   9.53 1022. 0.494
#> 17         3 GLMNET       Slice5       Resa…  313.  3.17 1.34   3.11  339. 0.918
#> 18         3 GLMNET       Slice6       Resa…  609.  6.03 2.02   6.26  667. 0.783