Calculate Accuracy Metrics from Modeltime Resamples

This is a wrapper for yardstick that simplifies time series regression accuracy metric calculations from a Modeltime Table that has been resampled and fitted using modeltime_fit_resamples().

Usage

modeltime_resample_accuracy(
  object,
  summary_fns = mean,
  metric_set = default_forecast_accuracy_metric_set(),
  ...
)

Arguments

object

a Modeltime Table with a column '.resample_results' (the output of modeltime_fit_resamples())

summary_fns

One or more functions to analyze resamples. The default is mean(). Possible values are:

NULL, to returns the resamples untransformed.

A function, e.g. mean.

A purrr-style lambda, e.g. ~ mean(.x, na.rm = TRUE)

A list of functions/lambdas, e.g. list(mean = mean, sd = sd)

metric_set

A yardstick::metric_set() that is used to summarize one or more forecast accuracy (regression) metrics.

...

Additional arguments passed to the function calls in summary_fns.

Details

#' Default Accuracy Metrics

The following accuracy metrics are included by default via modeltime::default_forecast_accuracy_metric_set():

MAE - Mean absolute error, yardstick::mae()

MAPE - Mean absolute percentage error, yardstick::mape()

MASE - Mean absolute scaled error, yardstick::mase()

SMAPE - Symmetric mean absolute percentage error, yardstick::smape()

RMSE - Root mean squared error, yardstick::rmse()

RSQ - R-squared, yardstick::rsq()

Summary Functions

By default, modeltime_resample_accuracy() returns the average accuracy metrics for each resample prediction.

The user can change this default behavior using summary_fns. Simply pass one or more Summary Functions. Internally, the functions are passed to dplyr::across(.fns), which applies the summary functions.

Returning Unsummarized Results

You can pass summary_fns = NULL to return unsummarized results by .resample_id.

Professional Tables (Interactive & Static)

Use modeltime::table_modeltime_accuracy() to format the results for reporting in reactable (interactive) or gt (static) formats, which are perfect for Shiny Apps (interactive) and PDF Reports (static).

Examples

library(modeltime)

# Mean (Default)
m750_training_resamples_fitted %>%
    modeltime_resample_accuracy() %>%
    table_modeltime_accuracy(.interactive = FALSE)
#> <div id="ssbocndpxt" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #ssbocndpxt .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #ssbocndpxt .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #ssbocndpxt .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #ssbocndpxt .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #ssbocndpxt .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #ssbocndpxt .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #ssbocndpxt .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #ssbocndpxt .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ssbocndpxt .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ssbocndpxt .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #ssbocndpxt .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #ssbocndpxt .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #ssbocndpxt .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #ssbocndpxt .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ssbocndpxt .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ssbocndpxt .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #ssbocndpxt .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #ssbocndpxt .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ssbocndpxt .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #ssbocndpxt .gt_left {
#>   text-align: left;
#> }
#> 
#> #ssbocndpxt .gt_center {
#>   text-align: center;
#> }
#> 
#> #ssbocndpxt .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #ssbocndpxt .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #ssbocndpxt .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #ssbocndpxt .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #ssbocndpxt .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #ssbocndpxt .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 65%;
#> }
#> </style>
#>   <table class="gt_table">
#>   <thead class="gt_header">
#>     <tr>
#>       <th colspan="10" class="gt_heading gt_title gt_font_normal gt_bottom_border" style>Accuracy Table</th>
#>     </tr>
#>     
#>   </thead>
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">.model_id</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">.model_desc</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">.type</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">n</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mae</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mape</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mase</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">smape</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">rmse</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">rsq</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_left">ARIMA(0,1,1)(0,1,1)[12]</td>
#> <td class="gt_row gt_left">Resamples</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_right">294.51</td>
#> <td class="gt_row gt_right">2.95</td>
#> <td class="gt_row gt_right">1.18</td>
#> <td class="gt_row gt_right">2.94</td>
#> <td class="gt_row gt_right">355.49</td>
#> <td class="gt_row gt_right">0.80</td></tr>
#>     <tr><td class="gt_row gt_right">2</td>
#> <td class="gt_row gt_left">PROPHET</td>
#> <td class="gt_row gt_left">Resamples</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_right">458.05</td>
#> <td class="gt_row gt_right">4.59</td>
#> <td class="gt_row gt_right">1.85</td>
#> <td class="gt_row gt_right">4.55</td>
#> <td class="gt_row gt_right">515.49</td>
#> <td class="gt_row gt_right">0.78</td></tr>
#>     <tr><td class="gt_row gt_right">3</td>
#> <td class="gt_row gt_left">GLMNET</td>
#> <td class="gt_row gt_left">Resamples</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_right">706.10</td>
#> <td class="gt_row gt_right">7.16</td>
#> <td class="gt_row gt_right">2.80</td>
#> <td class="gt_row gt_right">6.92</td>
#> <td class="gt_row gt_right">753.04</td>
#> <td class="gt_row gt_right">0.77</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

# Mean and Standard Deviation
m750_training_resamples_fitted %>%
    modeltime_resample_accuracy(
        summary_fns = list(mean = mean, sd = sd)
    ) %>%
    table_modeltime_accuracy(.interactive = FALSE)
#> <div id="ztpvgjeopf" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #ztpvgjeopf .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #ztpvgjeopf .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #ztpvgjeopf .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #ztpvgjeopf .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #ztpvgjeopf .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #ztpvgjeopf .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #ztpvgjeopf .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #ztpvgjeopf .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ztpvgjeopf .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ztpvgjeopf .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #ztpvgjeopf .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #ztpvgjeopf .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #ztpvgjeopf .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #ztpvgjeopf .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ztpvgjeopf .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ztpvgjeopf .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #ztpvgjeopf .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #ztpvgjeopf .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ztpvgjeopf .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #ztpvgjeopf .gt_left {
#>   text-align: left;
#> }
#> 
#> #ztpvgjeopf .gt_center {
#>   text-align: center;
#> }
#> 
#> #ztpvgjeopf .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #ztpvgjeopf .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #ztpvgjeopf .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #ztpvgjeopf .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #ztpvgjeopf .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #ztpvgjeopf .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 65%;
#> }
#> </style>
#>   <table class="gt_table">
#>   <thead class="gt_header">
#>     <tr>
#>       <th colspan="16" class="gt_heading gt_title gt_font_normal gt_bottom_border" style>Accuracy Table</th>
#>     </tr>
#>     
#>   </thead>
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">.model_id</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">.model_desc</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">.type</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">n</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mae_mean</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mae_sd</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mape_mean</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mape_sd</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mase_mean</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">mase_sd</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">smape_mean</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">smape_sd</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">rmse_mean</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">rmse_sd</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">rsq_mean</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">rsq_sd</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_left">ARIMA(0,1,1)(0,1,1)[12]</td>
#> <td class="gt_row gt_left">Resamples</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_right">294.51</td>
#> <td class="gt_row gt_right">121.98</td>
#> <td class="gt_row gt_right">2.95</td>
#> <td class="gt_row gt_right">1.27</td>
#> <td class="gt_row gt_right">1.18</td>
#> <td class="gt_row gt_right">0.56</td>
#> <td class="gt_row gt_right">2.94</td>
#> <td class="gt_row gt_right">1.24</td>
#> <td class="gt_row gt_right">355.49</td>
#> <td class="gt_row gt_right">138.68</td>
#> <td class="gt_row gt_right">0.80</td>
#> <td class="gt_row gt_right">0.15</td></tr>
#>     <tr><td class="gt_row gt_right">2</td>
#> <td class="gt_row gt_left">PROPHET</td>
#> <td class="gt_row gt_left">Resamples</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_right">458.05</td>
#> <td class="gt_row gt_right">196.09</td>
#> <td class="gt_row gt_right">4.59</td>
#> <td class="gt_row gt_right">1.98</td>
#> <td class="gt_row gt_right">1.85</td>
#> <td class="gt_row gt_right">0.91</td>
#> <td class="gt_row gt_right">4.55</td>
#> <td class="gt_row gt_right">1.86</td>
#> <td class="gt_row gt_right">515.49</td>
#> <td class="gt_row gt_right">175.26</td>
#> <td class="gt_row gt_right">0.78</td>
#> <td class="gt_row gt_right">0.09</td></tr>
#>     <tr><td class="gt_row gt_right">3</td>
#> <td class="gt_row gt_left">GLMNET</td>
#> <td class="gt_row gt_left">Resamples</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_right">706.10</td>
#> <td class="gt_row gt_right">248.29</td>
#> <td class="gt_row gt_right">7.16</td>
#> <td class="gt_row gt_right">2.62</td>
#> <td class="gt_row gt_right">2.80</td>
#> <td class="gt_row gt_right">1.11</td>
#> <td class="gt_row gt_right">6.92</td>
#> <td class="gt_row gt_right">2.41</td>
#> <td class="gt_row gt_right">753.04</td>
#> <td class="gt_row gt_right">252.75</td>
#> <td class="gt_row gt_right">0.77</td>
#> <td class="gt_row gt_right">0.16</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

# When summary_fns = NULL, returns the unsummarized resample results
m750_training_resamples_fitted %>%
    modeltime_resample_accuracy(
        summary_fns = NULL
    )
#> # A tibble: 18 × 10
#>    .model_id .model_desc  .resample_id .type   mae  mape  mase smape  rmse   rsq
#>        <int> <chr>        <chr>        <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1         1 ARIMA(0,1,1… Slice1       Resa…  182.  1.79 0.637  1.80  212. 0.950
#>  2         1 ARIMA(0,1,1… Slice2       Resa…  330.  3.24 1.30   3.17  406. 0.729
#>  3         1 ARIMA(0,1,1… Slice3       Resa…  359.  3.57 1.55   3.64  415. 0.859
#>  4         1 ARIMA(0,1,1… Slice4       Resa…  456.  4.76 2.01   4.61  530. 0.561
#>  5         1 ARIMA(0,1,1… Slice5       Resa…  122.  1.23 0.524  1.22  165. 0.926
#>  6         1 ARIMA(0,1,1… Slice6       Resa…  317.  3.12 1.05   3.20  407. 0.763
#>  7         2 PROPHET      Slice1       Resa…  450.  4.54 1.57   4.40  509. 0.869
#>  8         2 PROPHET      Slice2       Resa…  449.  4.34 1.76   4.45  494. 0.745
#>  9         2 PROPHET      Slice3       Resa…  475.  4.72 2.05   4.85  556. 0.660
#> 10         2 PROPHET      Slice4       Resa…  317.  3.33 1.39   3.26  362. 0.781
#> 11         2 PROPHET      Slice5       Resa…  813.  8.22 3.49   7.89  829. 0.897
#> 12         2 PROPHET      Slice6       Resa…  244.  2.40 0.808  2.45  344. 0.755
#> 13         3 GLMNET       Slice1       Resa…  893.  8.90 3.12   8.48  926. 0.845
#> 14         3 GLMNET       Slice2       Resa…  583.  5.73 2.29   5.53  647. 0.654
#> 15         3 GLMNET       Slice3       Resa…  874.  9.07 3.77   8.62  917. 0.901
#> 16         3 GLMNET       Slice4       Resa…  964. 10.1  4.24   9.53 1022. 0.494
#> 17         3 GLMNET       Slice5       Resa…  313.  3.17 1.34   3.11  339. 0.918
#> 18         3 GLMNET       Slice6       Resa…  609.  6.03 2.02   6.26  667. 0.783